% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinygouv-dependencies.R
\name{add_dsfr_deps}
\alias{add_dsfr_deps}
\title{ddsfr dependances}
\usage{
add_dsfr_deps(tag = NULL, version = get_dsfr_version())
}
\arguments{
\item{tag}{element de la page pour attacher les deps.}

\item{version}{ajout de la version pour les deps}
}
\description{
Ajout le css et le js de dsfr
}
