% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addwithspinner.R
\name{addWithSpinner}
\alias{addWithSpinner}
\title{spinner-specific html code}
\usage{
addWithSpinner(spinner_id = "lds-ring")
}
\arguments{
\item{spinner_id}{id du spinner}
}
\value{
an html code
}
\description{
spinner-specific html code
}
\examples{
addWithSpinner()
addWithSpinner()
}
