% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkboxinput_dsfr.R
\name{checkboxInput_dsfr}
\alias{checkboxInput_dsfr}
\title{checkboxInput_dsfr}
\usage{
checkboxInput_dsfr(inputId, label, value = FALSE)
}
\arguments{
\item{inputId}{id de l'input avec indexation}

\item{label}{id de l'input sans indexation}

\item{value}{TRUE (coché) ou FALSE (non coché)}
}
\value{
html
}
\description{
La case à cocher permet à l’utilisateur de sélectionner ou désélectionner une seule option. Elle est utilisée pour effectuer un choix binaire.
}
\examples{
checkboxInput_dsfr(
  inputId = 'test', label = 'test', value = FALSE)
checkboxInput_dsfr(
  inputId = 'test', label = 'test', value = TRUE
)
}
\seealso{
\code{\link[shiny:checkboxInput]{checkboxInput()}} pour en savoir plus sur la fonction Shiny originelle.
}
