% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_dsfr.R
\name{convert_to_dsfr}
\alias{convert_to_dsfr}
\title{convert_to_dsfr}
\usage{
convert_to_dsfr(path = "R/", version)
}
\arguments{
\item{path}{le chemin du dossier sur lequel faire la conversion, par defaut "R/"}

\item{version}{le numero de version du dsfr a utiliser, par exemple "1.7.2"}
}
\value{
des fichiers modifies
}
\description{
permet de convertir une application shiny en une application shiny suivant le DSFR
}
\examples{
mydir <- tempfile(pattern = "app")
dir.create(mydir)
golem::create_golem(
  mydir,
  overwrite = TRUE,
  open = FALSE
)
convert_to_dsfr(path = file.path(mydir, "R"), version = "1.7.2")
}
