% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daterangeinput_dsfr.R
\name{dateRangeInput_dsfr}
\alias{dateRangeInput_dsfr}
\title{dateRangeInput_dsfr}
\usage{
dateRangeInput_dsfr(inputId, label, start = NULL, end = NULL, separator = "to")
}
\arguments{
\item{inputId}{inputId}

\item{label}{label}

\item{start}{character La date de début au format aaaa-mm-jj. Si NULL (valeur par défaut), la date utilisée est la date du jour.}

\item{end}{character La date de fin au format aaaa-mm-jj. Si NULL (valeur par défaut), la date utilisée est la date du jour.}

\item{separator}{character Chaîne à afficher entre les zones de saisie de début et de fin de dates.}
}
\value{
html
}
\description{
dateRangeInput_dsfr
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {
  library(shiny)
  library(shinygouv)

  ui <- fluidPage_dsfr(
    header = header_dsfr(
      intitule = "Intitule",
      officiel = "Officiel",
      nom_site_service = "Nom du site / service",
      baseline = "baseline - precisions sur l organisation",
      class = "fr-m-1w"
    ),
    title = "Exemple",
    fluidRow_dsfr(
      # sans vecteur nommé
      dateRangeInput_dsfr(inputId = "daterange1",
                          label = "Date range:", start = "2001-01-01",separator = "à")
    )
  )
  server <- function(input, output, session) {

  observeEvent(input$daterange1, {
    print(input$daterange1)
  })

  }


  shinyApp(ui, server)
}
}
\seealso{
\code{\link[shiny:dateRangeInput]{dateRangeInput()}} in Shiny
}
