% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footer_bottom_dsfr.R, R/footer_dsfr.R,
%   R/footer_extra_url_bottom_dsfr.R
\name{footer_bottom_dsfr}
\alias{footer_bottom_dsfr}
\alias{footer_dsfr}
\alias{footer_extra_url_body_dsfr}
\alias{footer_extra_url_bottom_dsfr}
\alias{footer_li_dsfr}
\title{footer_bottom_dsfr}
\usage{
footer_bottom_dsfr(
  accessibilite = "non",
  ...,
  lien_accessibilite = "#",
  lien_mentions_legales = "#",
  lien_donnees_personnelles = "#",
  lien_gestion_cookies = "#"
)

footer_dsfr(
  intitule = NULL,
  description = NULL,
  accessibilite = "non",
  lien_accessibilite = "#",
  lien_mentions_legales = "#",
  lien_donnees_personnelles = "#",
  lien_gestion_cookies = "#",
  add_extra_url_body = NULL,
  add_extra_url_bottom = NULL,
  choix_theme = TRUE
)

footer_extra_url_body_dsfr(...)

footer_extra_url_bottom_dsfr(...)

footer_li_dsfr(id = NULL, href, titre = NULL, texte, type = "body")

footer_extra_url_bottom_dsfr(...)
}
\arguments{
\item{accessibilite}{le type d'accessibilité (non/partiellement/totalement conforme)}

\item{...}{d'autre liens de références (voir \code{footer_li_dsfr()})}

\item{lien_accessibilite}{lien url vers l'accessibilite}

\item{lien_mentions_legales}{lien url vers les mentions légales}

\item{lien_donnees_personnelles}{lien url vers les données personnelles}

\item{lien_gestion_cookies}{lien url vers la gestion des cookies}

\item{intitule}{la partie "intitule" du bloc marque. Utiliser un \code{shiny::tags$span()}pour mettre l'intitulé sur plusieurs lignes ou un vecteur.}

\item{description}{un texte de présentation - optionnel}

\item{add_extra_url_body}{url supplementaire a ajouter au corps du pied de page (voir \code{footer_extra_url_body_dsfr()})}

\item{add_extra_url_bottom}{url supplementaire a ajouter au bas du pied de page (voir \code{footer_extra_url_bottom_dsfr()})}

\item{choix_theme}{TRUE/FALSE (default) - affichage lien modal pour choix theme clair ou sombre dans le header}

\item{id}{un identifiant (optionnel)}

\item{href}{une url}

\item{titre}{le titre de l infobulle  (optionnel)}

\item{texte}{le texte}

\item{type}{le type de rendu de l url (body/bottom, par defaut body)}
}
\value{
html

html

html

html

html

html
}
\description{
Il se compose des éléments suivants:
\itemize{
\item la liste de liens liés aux obligations légales. Cette liste doit être définie en fonction du site, toutefois les liens & contenus suivants sont obligatoires : “accessibilité : non/partiellement/totalement conforme”, mentions légales, données personnelles et gestion des cookies. - obligatoire.
\item une mention de la licence - obligatoire “Sauf mention contraire, tous les contenus de ce site sont sous licence etalab-2.0”
}

Il est possible d'ajouter des url supplémentaires dans le corps et le bas du pied de page.

Permet d'ajouter des liends/url dans le corps du pied de page

Permet d'ajouter des liends/url dans le bas du pied de page

Permet de créer la balise html pour ajouter un lien en fonction de sa destination (body/bottom)

Permet d'ajouter des liends/url dans le bas du pied de page
}
\examples{
if (interactive()) {
  library(shiny)
  shinyApp(
    ui = fluidPage_dsfr(
      footer = footer_dsfr(
        intitule = c("Territoire", "Service"),
        description = "une description",
        accessibilite = "non",
        add_extra_url_body = footer_extra_url_body_dsfr(
          footer_li_dsfr(
            href = "https://siteinternet.fr",
            titre = "siteinternet.fr.fr - nouvelle fenêtre",
            texte = "siteinternet.fr.fr", type = "body"
          ),
          footer_li_dsfr(
            href = "https://siteinternet.fr",
            titre = "siteinternet.fr.fr - nouvelle fenêtre",
            texte = "siteinternet.fr.fr", type = "body"
          )
        ),
        add_extra_url_bottom = footer_extra_url_bottom_dsfr(

          footer_li_dsfr(
            href = "https://siteinternet.fr",
            titre = "siteinternet.fr.fr - nouvelle fenêtre",
            texte = "siteinternet.fr.fr", type = "bottom"
          )
        )
      ),
      tabsetPanel_dsfr(
        "coucou",
        tabPanel_dsfr(
          id = "tab1",
          title = "Une prems tab",
          content = tagList(h3("Coucou"),
                            plotOutput("plot"))
        ),
        tabPanel_dsfr(
          id = "tab2",
          title = "Une deuxieme tab",
          content = tagList(h3("Salut"),
                            plotOutput("plot2"))
        )
      )
    ),
    server = function(input, output) {
      output$plot <- renderPlot({
        plot(cars)
      })
      output$plot2 <- renderPlot({
        plot(iris)
      })
    }
  )
}
footer_extra_url_body_dsfr(
                footer_li_dsfr(href = "http://siteinternet.fr", texte = "siteinternet.fr", type = "body"),
                footer_li_dsfr(href = "http://siteinternet.fr", texte = "siteinternet.fr", type = "body"))
footer_extra_url_bottom_dsfr(footer_li_dsfr(href = "http://siteinternet.fr", texte = "siteinternet.fr"))
footer_li_dsfr(href = "http://siteinternet.fr", texte = "siteinternet.fr", type = "body")
footer_li_dsfr(href = "http://siteinternet.fr", texte = "siteinternet.fr", type = "bottom")
}
