% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_dsfr.R
\name{header_dsfr}
\alias{header_dsfr}
\title{header_dsfr
Vous pouvez ajouter le header officiel dans votre application avec \code{header_dsfr()}.}
\usage{
header_dsfr(
  intitule = c("intitulé", "du", "bloc-marque"),
  url = "/",
  titre = NULL,
  nom_site_service = NULL,
  baseline = NULL,
  class = "fr-container",
  choix_theme = FALSE,
  logo = NULL,
  fastlink = NULL
)
}
\arguments{
\item{intitule}{la partie "intitule" du bloc marque. Utiliser un \code{shiny::tags$span()}pour mettre l'intitulé sur plusieurs lignes ou un vecteur.}

\item{url}{l url de redirection - optionnel}

\item{titre}{le texte de description de l url - optionnel}

\item{nom_site_service}{le nom du site ou du service - optionnel}

\item{baseline}{le texte sous le nom de site/service - optionnel}

\item{class}{un classe a ajouter si necessaire ("fr-m-1w" correspond a une margin de 1w sur tous les cotés) - optionnel, par defaut "fr-container" (centré) (voir le Systeme de design de l'Etat "fondamentaux-techniques/espacements")}

\item{choix_theme}{TRUE/FALSE (default) - affichage lien modal pour choix theme clair ou sombre dans le header}

\item{logo}{list(url,alttext,style) - affichage d'un logo dans le header à droite de la marianne. Les attributs doivent être tous remplis si on souhaite un logo - optionnel}

\item{fastlink}{list(list(lib, url, icon, blank)) - affichage de lien rapide dans le header. Les attributs doivent être tous remplis si on souhaite des liens rapide - optionnel (pour les icones voir https://www.systeme-de-design.gouv.fr/fondamentaux/icone/)}
}
\value{
html
}
\description{
header_dsfr
Vous pouvez ajouter le header officiel dans votre application avec \code{header_dsfr()}.
}
\examples{
if(interactive()) {
 shiny::shinyApp(
  ui = fluidPage_dsfr(
    header = header_dsfr(
      intitule = c("Pr\u00e9fet", "de", "Bretagne"),
      fastlink = list(list(lib="spyrales/shinygouv", url="https://github.com/spyrales/shinygouv/", icon="fr-icon-github-fill", blank=FALSE))
    )
  ),
    server = function(input, output) {}
 )
}
}
