% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidespinner_dsfr.R
\name{hideSpinner_dsfr}
\alias{hideSpinner_dsfr}
\title{Masque le spinner}
\usage{
hideSpinner_dsfr(
  spinner_id = "lds-ring",
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{spinner_id}{id du spinner}

\item{session}{session}
}
\value{
html
}
\description{
Masque le spinner dans le cadre d'un appel à `shinygouv::withSpinner_dsfr()`
}
\examples{
\dontrun{
hideSpinner_dsfr()
}
}
