% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modals.R
\name{modalDialog_dsfr}
\alias{modalDialog_dsfr}
\alias{showModal_dsfr}
\alias{removeModal_dsfr}
\title{modalDialog_dsfr}
\usage{
modalDialog_dsfr(..., inputId = "fr-modal-1", title, footer = NULL)

showModal_dsfr(
  inputId = "fr-modal-1",
  session = shiny::getDefaultReactiveDomain()
)

removeModal_dsfr(
  inputId = "fr-modal-1",
  session = shiny::getDefaultReactiveDomain()
)
}
\value{
html
}
\description{
modalDialog_dsfr
}
\examples{

modalDialog_dsfr(
  htmltools::div("Le corps du texte"),
  title = "Titre du modal",
  footer = actionButton_dsfr(inputId = "Bouton",
                             label = "Bouton")
)


}
