% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbarpage_dsfr.R
\name{navbarPanel_dsfr}
\alias{navbarPanel_dsfr}
\title{Panel pour la navbar}
\usage{
navbarPanel_dsfr(title, ..., class = "fr-container")
}
\arguments{
\item{title}{Titre du panel}

\item{...}{Contenu du panel}

\item{class}{Classe CSS du panel}
}
\description{
Panel pour la navbar
}
