% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_html.R
\name{parse_html}
\alias{parse_html}
\title{Traduit du html en resultat shiny}
\usage{
parse_html(html_code, zone = "/html/body/*")
}
\arguments{
\item{html_code}{resultat d'une fonction htmlTemplate ou une chaine de caractere}

\item{zone}{zone a garder pour parser}
}
\value{
shiny.tag
}
\description{
Traduit du html en resultat shiny
}
\examples{
html_code <- '<button id="test" type="button" class="btn btn-default action-button">test</button>'
parse_html(html_code)
}
