% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_node.R
\name{render_node}
\alias{render_node}
\title{render_node}
\usage{
render_node(node, indent = 0, prefix = FALSE)
}
\arguments{
\item{node}{node qui provient de parse_to_xml}

\item{indent}{ajout de l'indentation}

\item{prefix}{si TRUE alors le resultat sera prefixe par tags}
}
\description{
render_node
}
