% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton.R
\name{shinygouv_file}
\alias{shinygouv_file}
\title{Trouve les chemins des fichiers du package shinygouv}
\usage{
shinygouv_file(...)
}
\arguments{
\item{...}{Une ou plusieurs chaines de caracteres, representant les noms de fichiers ou repertoires a rechercher dans le package.}
}
\value{
Un chemin absolu.
}
\description{
Renvoie les adresses absolues des fichiers presents dans le repertoire d'installation du package a partir de leur adresse relative.
}
\examples{
shinygouv_file('rstudio', 'templates', 'project', 'ressources')

}
