% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showspinner_dsfr.R
\name{showSpinner_dsfr}
\alias{showSpinner_dsfr}
\title{Affiche le spinner}
\usage{
showSpinner_dsfr(
  spinner_id = "lds-ring",
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{spinner_id}{id du spinner}

\item{session}{session}
}
\value{
html
}
\description{
Affiche le spinner dans le cadre d'un appel à `shinygouv::withSpinner_dsfr()`
}
\examples{
\dontrun{
showSpinner_dsfr()
}
}
