% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebarlayout_dsfr.R
\name{sidebarLayout_dsfr}
\alias{sidebarLayout_dsfr}
\alias{sidebarPanel_dsfr}
\alias{mainPanel_dsfr}
\title{sidebarLayout_dsfr}
\usage{
sidebarLayout_dsfr(sidebarPanel, mainPanel, position = c("left", "right"))

sidebarPanel_dsfr(..., width = 4, title = NULL)

mainPanel_dsfr(..., width = 8)
}
\arguments{
\item{sidebarPanel}{Le \code{sidebarPanel_dsfr()} contenant les contrôles d'entrée.}

\item{mainPanel}{Le \code{mainPanel_dsfr()} contenant les sorties.}

\item{position}{La position de la barre latérale par rapport à la zone principale ("left"
ou "right").}

\item{...}{Éléments de sortie à inclure dans la barre latérale/panneau principal.}

\item{width}{La largeur de la barre latérale et du panneau principal. Par défaut,
La barre latérale occupe 1/3 de la largeur et le panneau principal 2/3. Le total
doit être inférieur ou égale à 12.}

\item{title}{titre du panneau de barre latérale (facultatif)}
}
\value{
html
}
\description{
Créez une mise en page (\code{sidebarLayout_dsfr()}) avec une barre latérale (\code{sidebarPanel_dsfr()}) et
une zone principale (\code{mainPanel_dsfr()}). La barre latérale peut s'afficher avec une
couleur d'arrière-plan grise et contient généralement des contrôles d'entrée.
La zone principale occupe les 2/3 de la largeur et contient généralement des sorties.
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {
options(device.ask.default = FALSE)

# Define UI
ui <- fluidPage_dsfr(

  # Application title
  titlePanel("Hello Shiny!"),

  sidebarLayout_dsfr(

    # Sidebar with a slider input
    sidebarPanel_dsfr(
      sliderInput_dsfr("obs",
                  "Number of observations:",
                  min = 0,
                  max = 1000,
                  value = 500),
      width = 4,
      title = "My sidebar"
    ),

    # Show a plot of the generated distribution
    mainPanel_dsfr(
      plotOutput("distPlot")
    )
  )
)

# Server logic
server <- function(input, output) {
  output$distPlot <- renderPlot({
    hist(rnorm(input$obs))
  })
}

# Complete app with UI and server components
shinyApp(ui, server)
}
}
\references{
inspiré de la fonction \code{\link[shiny:sidebarLayout]{shiny::sidebarLayout()}}
}
\seealso{
Other layout functions:
\code{\link[=wellPanel_dsfr]{wellPanel_dsfr()}}
}
\concept{layout functions}
