% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableoutput_dsfr.R
\name{tableOutput_dsfr}
\alias{tableOutput_dsfr}
\title{tableOutput_dsfr}
\usage{
tableOutput_dsfr(outputId)
}
\arguments{
\item{outputId}{nom d'output a partir duquel lire la table}
}
\value{
html
}
\description{
tableOutput_dsfr
}
\examples{

if (interactive()) {
  # table example
  shinyApp(
    ui = fluidPage_dsfr(
      fluidRow_dsfr(
        column_dsfr(12,
          tableOutput_dsfr('table')
        )
      )
    ),
    server = function(input, output) {
      output$table <- renderTable(iris)
    }
  )
}
}
\seealso{
\code{\link[shiny:tableOutput]{tableOutput()}}
}
