% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabpanel.R
\name{tabsetPanel_dsfr}
\alias{tabsetPanel_dsfr}
\alias{tabPanel_dsfr}
\title{tabsetPanel_dsfr}
\usage{
tabsetPanel_dsfr(id, ...)

tabPanel_dsfr(id, title, content)
}
\arguments{
\item{id}{id du tabset/du tabPanel}

\item{...}{liste de tabPanel_dsfr}

\item{title}{titre du tabset}

\item{content}{contenu du tabset}
}
\value{
html
}
\description{
tabsetPanel_dsfr
}
\examples{
if (interactive()){
  library(shiny)
  shinyApp(
    ui = fluidPage_dsfr(
      tabsetPanel_dsfr(
        "coucou",
        tabPanel_dsfr(
          id = "tab1",
          title = "Une prems tab",
          content = tagList(
            h3("Coucou"),
            plotOutput("plot")
          )
        ),
        tabPanel_dsfr(
          id = "tab2",
          title = "Une deuxieme tab",
          content = tagList(
            h3("Salut"),
            plotOutput("plot2")
          )
        )
      )
    ),
    server = function(input, output) {
      output$plot <- renderPlot({
        plot(cars)
      })
      output$plot2 <- renderPlot({
        plot(iris)
      })
    }
  )
}
}
\seealso{
\code{\link[shiny:tabsetPanel]{tabsetPanel()}} in Shiny

\code{\link[shiny:tabPanel]{tabPanel()}} in Shiny
}
