% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggleswitch_dsfr.R
\name{toggleSwitch_dsfr}
\alias{toggleSwitch_dsfr}
\title{toggleSwitch_dsfr}
\usage{
toggleSwitch_dsfr(inputId, label, activate = "", deactivate = "")
}
\arguments{
\item{inputId}{id de l'input}

\item{label}{label du bouton}

\item{activate}{sous texte quand activé}

\item{deactivate}{sous texte quand desactivé}
}
\value{
html
}
\description{
toggleSwitch_dsfr
}
\examples{
if (interactive()) {
  library(shiny)
  ui <- fluidPage_dsfr(
    toggleSwitch_dsfr(inputId = "test", label = "test")
  )

  server <- function(input, output, session) {
    observeEvent(input$test, {
      message(input$test)
      message("coucou")
    })
  }

  shinyApp(ui, server)
}
}
