% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatecheckboxgroupinput_dsfr.R
\name{updateCheckboxGroupInput_dsfr}
\alias{updateCheckboxGroupInput_dsfr}
\title{updateCheckboxGroupInput_dsfr}
\usage{
updateCheckboxGroupInput_dsfr(
  inputId,
  label = NULL,
  choices = NULL,
  selected = NULL,
  inline = FALSE,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{inputId}{id de l'input}

\item{label}{label du bouton}

\item{choices}{Liste des valeurs à sélectionner (si les éléments de la liste portent un nom, c'est ce nom qui est affiché à l'utilisateur et non la valeur)}

\item{selected}{Valeurs préselectionnées}

\item{inline}{Si TRUE, positionne les choix en ligne (c'est-à-dire horizontalement).}

\item{session}{la session, la valeur par défaut est getDefaultReactiveDomain().}
}
\value{
html
}
\description{
updateCheckboxGroupInput_dsfr
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {
  
  library(shiny)
  
  ui <- fluidPage_dsfr(
    checkboxGroupInput_dsfr(
      "variable", "Variables to show:",
      c("Cylinders" = "cyl",
        "Transmission" = "am",
        "Gears" = "gear"),
      inline = FALSE),
    actionButton_dsfr("go", "Change for inline and choices"),
    actionButton_dsfr("go2", "Change without inline and choices")
  )
  
  server <- function(input, output, session) {
    observeEvent(input$variable, {
      print(input$variable)
    })
    
    
    observeEvent(input$go, {
      updateCheckboxGroupInput_dsfr(
        session = session,
        inputId = "variable",
        label = paste0("test", rnorm(1)),
        choices = c("A" = "a", "B" = "b"),
        selected = "a",
        inline = TRUE
      )
      
    })
    
    
    observeEvent(input$go2, {
      updateCheckboxGroupInput_dsfr(
        session = session,
        inputId = "variable",
        label = paste0("test", rnorm(1)),
        choices = c("A" = "a", "B" = "b"),
        selected = "a",
        inline = FALSE
      )
      
    })
    
  }
  shinyApp(ui, server)
}
}
\seealso{
\code{\link[shiny:updateCheckboxInput]{updateCheckboxInput()}} pour en savoir plus sur la fonction originelle de Shiny.
}
