% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatecheckboxinput_dsfr.R
\name{updateCheckboxInput_dsfr}
\alias{updateCheckboxInput_dsfr}
\title{updateCheckboxInput_dsfr}
\usage{
updateCheckboxInput_dsfr(
  inputId,
  label = NULL,
  value = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{inputId}{id de l'input}

\item{label}{label du bouton}

\item{value}{TRUE (coché) ou FALSE (non coché)}

\item{session}{la session, la valeur par défaut est getDefaultReactiveDomain().}
}
\value{
html
}
\description{
Change la valeur d'un checkboxInput
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {
  ui <- fluidPage_dsfr(
    checkboxInput_dsfr(inputId = 'inCheckboxInput', label = 'test', value = FALSE),
    actionButton_dsfr("go", "Change")
  )
  server <- function(input, output, session) {
    
    observeEvent(input$go, {
      updateCheckboxInput_dsfr(
        session = session,
        inputId = "inCheckboxInput",
        label = "Un nouveau label",
        value = TRUE
      )
    })
    
    observeEvent(input$inCheckboxInput, {
      message(input$inCheckboxInput)
    })
  }
  shinyApp(ui, server)
}
}
\seealso{
\code{\link[shiny:updateCheckboxInput]{updateCheckboxInput()}} pour en savoir plus sur la fonction Shiny originelle.
}
