% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_favicon_dsfr.R
\name{use_favicon_dsfr}
\alias{use_favicon_dsfr}
\title{use_favicon_dsfr}
\usage{
use_favicon_dsfr(path_app = "", golem_app = TRUE)
}
\arguments{
\item{path_app}{chaine de texte indiquant le chemin vers votre application shiny}

\item{golem_app}{booleen, TRUE pour une app golem, FALSE sinon}
}
\description{
use_favicon_dsfr() est une fonction a lancer dans la console pour enregistrer le fichier de favicon marianne dans votre projet shiny.
}
\examples{
\dontrun{
# A lancer dans la console pour rapatrier le fichier dans votre repertoire de travail
use_favicon_dsfr(path_app = "", golem_app = TRUE)
}
}
