% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellpanel_dsfr.R
\name{wellPanel_dsfr}
\alias{wellPanel_dsfr}
\title{wellPanel_dsfr}
\usage{
wellPanel_dsfr(..., title = NULL, grey = FALSE)
}
\arguments{
\item{...}{Éléments d'interface utilisateur à inclure à l'intérieur du panneau.}

\item{title}{titre du panneau (optionnel)}

\item{grey}{arrière-plan gris du panneau (FALSE par défaut)}
}
\value{
html
}
\description{
Crée un panneau avec une bordure légèrement incrustée et un arrière-plan gris et un titre, si on le souhaite.
}
\examples{
if (interactive()) {
  shiny::shinyApp(
    ui = fluidPage_dsfr(
      title = "exemple",
      fluidRow_dsfr(
      column_dsfr(
        12,
        wellPanel_dsfr("Ceci est un wellPanel avec titre", title="Mon wellPanel"),
        wellPanel_dsfr("Ceci est un wellPanel avec option grey et sans titre", grey = TRUE)
      )
      )
    ),
    server = function(input, output) {}
  )
}
}
\references{
inspiré de la fonction \code{\link[shiny:wellPanel]{shiny::wellPanel()}}
}
\seealso{
Other layout functions:
\code{\link[=sidebarLayout_dsfr]{sidebarLayout_dsfr()}}
}
\concept{layout functions}
