% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actionbutton_dsfr.R
\name{actionButton_dsfr}
\alias{actionButton_dsfr}
\title{actionButton_dsfr}
\usage{
actionButton_dsfr(inputId, label, class = "fr-p-1w")
}
\arguments{
\item{inputId}{id de l'input}

\item{label}{label du bouton}

\item{class}{des classes a ajouter si necessaire}
}
\value{
html
}
\description{
actionButton_dsfr
}
\examples{
if (interactive()) {
  library(shiny)
  library(shinygouv)

  ui <- fluidPage_dsfr(
    header = header_dsfr(
      intitule = "Intitule",
      officiel = "Officiel",
      nom_site_service = "Nom du site / service",
      baseline = "baseline - precisions sur l organisation",
      class = "fr-m-1w"
    ),
    title = "Exemple",
    fluidRow_dsfr(
      actionButton_dsfr(
        inputId = "test",
        label = "Cliquez moi",
        class = "fr-btn--icon-left fr-icon-checkbox-circle-line"
      ),
      verbatimTextOutput(outputId = "nbclic")
    )
  )
  server <- function(input, output, session) {
    output$nbclic <- renderText({
      paste("Vous avez cliqu\u00e9", input$test, "fois sur le bouton")
    })
  }

  shinyApp(ui, server)
}
}
\seealso{
\code{\link[shiny:actionButton]{shiny::actionButton()}} pour en savoir plus sur la création de boutons d'action avec Shiny.
}
