% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actionbutton_dsfr.R
\name{actionButton_dsfr}
\alias{actionButton_dsfr}
\title{actionButton_dsfr}
\usage{
actionButton_dsfr(inputId, label, class = NULL)
}
\arguments{
\item{inputId}{id de l'input}

\item{label}{label du bouton}

\item{class}{des classes a ajouter si necessaire}
}
\value{
html
}
\description{
actionButton_dsfr
}
\examples{
actionButton_dsfr(
  inputId = "test",
  label = "test",
  class = "fr-btn--icon-left fr-icon-checkbox-circle-line"
)
}
