% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinygouv-dependencies.R
\name{add_dsfr_deps}
\alias{add_dsfr_deps}
\title{ddsfr dependances}
\usage{
add_dsfr_deps(tag)
}
\arguments{
\item{tag}{element de la page pour attacher les deps.}
}
\description{
Ajout le css et le js de dsfr
}
