% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkboxgroupinput_dsfr.R
\name{checkboxGroupInput_dsfr}
\alias{checkboxGroupInput_dsfr}
\title{checkboxGroupInput_dsfr}
\usage{
checkboxGroupInput_dsfr(
  inputId,
  label,
  choices = NULL,
  selected = NULL,
  inline = FALSE
)
}
\arguments{
\item{inputId}{id de l'input}

\item{label}{label du bouton}

\item{choices}{Liste des valeurs à sélectionner (si les éléments de la liste portent un nom, c'est ce nom qui est affiché à l'utilisateur et non la valeur)}

\item{selected}{Valeurs préselectionnées}

\item{inline}{Si TRUE, positionne les choix en ligne (c'est-à-dire horizontalement).}
}
\value{
html
}
\description{
checkboxGroupInput_dsfr
}
\examples{
if (interactive()) {
  
  library(shiny)
  
  ui <- fluidPage_dsfr(
    checkboxGroupInput_dsfr("checkboxes", "Variables to show:",
                            c("Cylinders" = "cyl",
                              "Transmission" = "am",
                              "Gears" = "gear"), inline = TRUE)
  )
  
  server <- function(input, output, session) {
    observeEvent(input$checkboxes, {
      print(input$checkboxes)
    })
  }
  
  shinyApp(ui, server)
}
}
\seealso{
\code{\link[shiny:checkboxGroupInput]{shiny::checkboxGroupInput()}} pour en savoir plus sur la fonction originelle de Shiny.
}
