% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkboxgroupinput_dsfr.R
\name{checkboxGroupInput_dsfr}
\alias{checkboxGroupInput_dsfr}
\title{checkboxGroupInput_dsfr}
\usage{
checkboxGroupInput_dsfr(inputId, label, choices = NULL, selected = NULL)
}
\arguments{
\item{inputId}{id de l'input}

\item{label}{label du bouton}

\item{choices}{Liste des valeurs à sélectionner (si les éléments de la liste portent un nom, c'est ce nom qui est affiché à l'utilisateur et non la valeur)}

\item{selected}{Valeurs préselectionnées}
}
\value{
html
}
\description{
checkboxGroupInput_dsfr
}
\examples{

if (interactive()) {
  library(shiny)

  ui <- fluidPage_dsfr(
    checkboxGroupInput_dsfr(
      "variable",
      "Variables to show:",
      c(
        "Cylinders" = "cyl",
        "Transmission" = "am",
        "Gears" = "gear"
      )
    )
  )

  server <- function(input, output, session) {
    observeEvent(input$variable, {
      print(input$variable)
    })
  }

  shinyApp(ui, server)
}

}
