% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkboxinput_dsfr.R
\name{checkboxInput_dsfr}
\alias{checkboxInput_dsfr}
\title{checkboxInput_dsfr}
\usage{
checkboxInput_dsfr(
  inputId,
  label,
  choices = NULL,
  selected = NULL,
  inline = FALSE,
  choice_names = NULL,
  choice_values = NULL
)
}
\arguments{
\item{inputId}{id de l'input}

\item{label}{label du bouton}

\item{choices}{Liste des valeurs à sélectionner (si les éléments de la liste portent un nom, c'est ce nom qui est affiché à l'utilisateur et non la valeur)}

\item{selected}{Valeurs préselectionnées}

\item{inline}{Si TRUE, positionne les choix en ligne (c'est-à-dire horizontalement).}

\item{choice_names}{Liste de labels (ignoré si choices est défini)}

\item{choice_values}{Liste des valeurs (ignoré si choices est défini)}
}
\value{
html
}
\description{
checkboxInput_dsfr
}
