% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_dsfr.R
\name{column_dsfr}
\alias{column_dsfr}
\title{column_dsfr}
\usage{
column_dsfr(width = NULL, ...)
}
\arguments{
\item{width}{width par défaut est NULL, cela signifie que la taille de la colonne depend des autres colonnes. Vous pouvez aussi utiliser 0}

\item{...}{differents tag pour construire la page}
}
\value{
shiny.tag.list
}
\description{
column_dsfr
}
\examples{

if (interactive()) {
  shiny::shinyApp(
    ui = fluidPage_dsfr(
      title = "exemple",
      fluidRow_dsfr(
        column_dsfr(0, "test"),
        column_dsfr(0, "test"),
        column_dsfr(0, "test"),
        column_dsfr(0, "test")
      )
    ),
    server = function(input, output) {}
  )
}

}
