% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_dsfr.R
\name{convert_to_dsfr}
\alias{convert_to_dsfr}
\title{convert_to_dsfr}
\usage{
convert_to_dsfr(path = "R/", version = get_dsfr_version())
}
\arguments{
\item{path}{le chemin du dossier sur lequel faire la conversion, par defaut "R/"}

\item{version}{le numero de version du dsfr a utiliser, par exemple "1.7.2". Par défaut, la derniere disponible. Utilisé plutot lors du developpement pour tester les retrocompatibilites.}
}
\value{
des fichiers modifies
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Il s'agissait d'une fonction expérimentale qui vous permet de convertir une application shiny en une application shiny suivant le DSFR. Il reste des pionts de vigilances sur la convertion des tabPanel dans les navbarPage et les tabsetPanel
}
\examples{
if (FALSE) {
  mydir <- tempfile(pattern = "app")
  dir.create(mydir)
  golem::install_dev_deps(
    force_install = TRUE
  )
  golem::create_golem(
    mydir,
    overwrite = TRUE,
    open = FALSE
  )
  convert_to_dsfr(path = file.path(mydir, "R"))
}
}
