% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileinput_dsfr.R
\name{fileInput_dsfr}
\alias{fileInput_dsfr}
\title{fileInput_dsfr}
\usage{
fileInput_dsfr(inputId, label, message)
}
\arguments{
\item{inputId}{id de l'input}

\item{label}{etiquette du bouton d'import, par exemple "choisir fichier"}

\item{message}{eventuelles indications complementaires, par exemple "fichier images de moins de 2 Mo"}
}
\value{
html
}
\description{
fileInput_dsfr
}
\examples{

if (interactive()) {
  ui <- fluidPage_dsfr(
    fluidRow_dsfr(
      column_dsfr(width = 4,
        fileInput_dsfr(inputId = "file1", label = "Importez vos données", message = "seul le format csv est pris en charge"),
        br(),
        checkboxInput_dsfr(inputId = "header", label = "Utiliser les en-tetes du fichier")
      ),
      column_dsfr(width = 8,
        tableOutput_dsfr("contenu")("contenu")
      )
    )
  )

  server <- function(input, output) {
    output$contenu <- renderTable({
      file <- input$file1
      ext <- tools::file_ext(file$datapath)

      req(file)
      validate(need(ext == "csv", "Veuillez choisir un fichier csv svp"))

      read.csv2(file$datapath, header = input$header, nrows = 12)
    })
  }

  shinyApp(ui, server)
}


}
\seealso{
\code{\link[=fileInput]{fileInput()}} Pour en savoir plus sur la fonction originelle de Shiny.
\href{https://shiny.posit.co/r/reference/shiny/latest/fileinput}{Documentation externe}
}
