% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluidpage_dsfr.R
\name{fluidPage_dsfr}
\alias{fluidPage_dsfr}
\title{fluidPage_dsfr}
\usage{
fluidPage_dsfr(
  ...,
  header = NULL,
  title = NULL,
  footer = NULL,
  theme = NULL,
  lang = NULL,
  dependencies = NULL
)
}
\arguments{
\item{...}{element a inclure dans la page}

\item{header}{l entete de la page (voir \code{header_dsfr()})}

\item{title}{titre de la page}

\item{footer}{contenu du pied de page (voir \code{footer_dsfr()})}

\item{theme}{pas implemente}

\item{lang}{pas implemente}

\item{dependencies}{dépendences complémentaire à implémenter
dans la balise head (exemple golem add_resource_path ou htmlDependency).
NULL par défaut}
}
\value{
html
}
\description{
fluidPage_dsfr
}
\examples{
if (interactive()) {

  library(shiny)
  my_page <- fluidPage_dsfr(
    header = header_dsfr(
      intitule = c("Prefet", "de", "Bretagne")
    ),
    title = "Gouv",
    htmltools::div("test")
  )

  shiny::shinyApp(
    my_page,
    server = function(input, output) {}
  )
}
}
\references{
inspiré de la fonction \code{\link[shiny:fluidPage]{shiny::fluidPage()}}
}
\seealso{
Other structure functions:
\code{\link[=navbarPage_dsfr]{navbarPage_dsfr()}}
}
\concept{structure functions}
