% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluidpage_dsfr.R
\name{fluidPage_dsfr}
\alias{fluidPage_dsfr}
\title{fluidPage_dsfr}
\usage{
fluidPage_dsfr(..., header = NULL, title = NULL, theme = NULL, lang = NULL)
}
\arguments{
\item{...}{element a inclure dans la page}

\item{header}{l entete de la page}

\item{title}{titre de la page}

\item{theme}{pas implemente}

\item{lang}{pas implemente}
}
\value{
html
}
\description{
fluidPage_dsfr
}
\examples{
my_page <- fluidPage_dsfr(
  header = header_dsfr(
    intitule = "Prefet de",
    officiel = "Bretagne",
  ),
  title = "Gouv",
  htmltools::div("test")
)
if (interactive()) {
  shiny::shinyApp(
    my_page,
    server = function(input, output) {}
  )
}
}
