% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluidrow_dsfr.R
\name{fluidRow_dsfr}
\alias{fluidRow_dsfr}
\title{fluidRow_dsfr}
\usage{
fluidRow_dsfr(..., class = NULL)
}
\arguments{
\item{...}{differents tag pour construire la page}

\item{class}{ajout de class pour la fluidRow_dsfr (voir la page officiel du CSS)}
}
\value{
shiny.tag.list
}
\description{
fluidRow_dsfr
}
\examples{
if (interactive()) {
  shiny::shinyApp(
    ui = fluidPage_dsfr(
      title = "exemple",
      fluidRow_dsfr(
        shiny::p("Voici un exemple")
      )
    ),
    server = function(input, output) {}
  )
}
}
\seealso{
\code{\link[shiny:fluidPage]{shiny::fluidRow()}} fonction shiny originelle
}
