% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dsfr_version.R
\name{get_dsfr_version}
\alias{get_dsfr_version}
\title{Recuperer la version du dsfr de 'shinygouv'}
\usage{
get_dsfr_version()
}
\value{
Character. Numero de version
}
\description{
Recuperer la version du dsfr de 'shinygouv'
}
\examples{
get_dsfr_version()
}
