% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_dsfr.R
\name{header_dsfr}
\alias{header_dsfr}
\title{header_dsfr}
\usage{
header_dsfr(
  intitule,
  officiel,
  url = "/",
  titre = NULL,
  nom_site_service = NULL,
  baseline = NULL,
  class = "fr-container"
)
}
\arguments{
\item{intitule}{la partie "intitule" du bloc marque}

\item{officiel}{la partie "officiel" du bloc marque}

\item{url}{l url de redirection - optionnel}

\item{titre}{le texte de description de l url - optionnel}

\item{nom_site_service}{le nom du site ou du service - optionnel}

\item{baseline}{le texte sous le nom de site/service - optionnel}

\item{class}{un classe a ajouter si necessaire ("fr-m-1w" correspond a une margin de 1w sur tous les cotés) - optionnel, par defaut "fr-container" (centré) (voir le Systeme de design de l'Etat "fondamentaux-techniques/espacements")}
}
\value{
html
}
\description{
header_dsfr
}
\examples{

header_dsfr(
  intitule = "Prefet de",
  officiel = "Bretagne",
)
}
