% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modaldialog_dsfr.R
\name{modalDialog_dsfr}
\alias{modalDialog_dsfr}
\title{modalDialog_dsfr}
\usage{
modalDialog_dsfr(..., inputId = "fr-modal-1", title, footer = NULL)
}
\arguments{
\item{...}{element a inclure dans le modal}

\item{inputId}{inputId du modal}

\item{title}{titre du modal}

\item{footer}{footer du modal}
}
\value{
html
}
\description{
modalDialog_dsfr
}
\examples{
if (interactive()) {
  library(shiny)
  ui <- fluidPage_dsfr(
    actionButton_dsfr("show", "Show modal dialog")
  )
  server <- function(input, output, session) {
    observeEvent(input$show, {
      showModal_dsfr(
        ui = modalDialog_dsfr(
          tagList(
            htmltools::div("Le corps du texte"),
            actionButton_dsfr(
              inputId = "btn",
              label = "txt"
            )
          ),
          title = "Titre du modal",
          footer = "Footer du modal"
        )
      )
    })

    observe({
      print(input$btn)
    })
  }
  shinyApp(ui, server)
}
}
\seealso{
\code{\link[shiny:modalDialog]{shiny::modalDialog()}} in Shiny
}
