% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modaldialog_dsfr.R
\name{modalDialog_dsfr}
\alias{modalDialog_dsfr}
\title{modalDialog_dsfr}
\usage{
modalDialog_dsfr(..., inputId = "fr-modal-1", title, footer = NULL)
}
\arguments{
\item{...}{element a inclure dans le modal}

\item{inputId}{inputId du modal}

\item{title}{titre du modal}

\item{footer}{footer du modal}
}
\value{
html
}
\description{
modalDialog_dsfr
}
\examples{

modalDialog_dsfr(
  htmltools::div("Le corps du texte"),
  title = "Titre du modal",
  footer = actionButton_dsfr(inputId = "Bouton",
                             label = "Bouton")
)


}
