% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericinput_dsfr.R
\name{numericInput_dsfr}
\alias{numericInput_dsfr}
\title{numericInput_dsfr}
\usage{
numericInput_dsfr(
  inputId = inputId,
  label = label,
  value = value,
  min = NA,
  max = NA,
  step = NA
)
}
\arguments{
\item{inputId}{inputId}

\item{label}{label}

\item{value}{valeur par défaut}

\item{min}{valeur minimale}

\item{max}{valeur maximale}

\item{step}{pas utilisé entre 2 valeurs}
}
\value{
html
}
\description{
numericInput_dsfr
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {
  library(shiny)
  library(shinygouv)

  ui <- fluidPage_dsfr(
    header = header_dsfr(
      intitule = "Intitule",
      officiel = "Officiel",
      nom_site_service = "Nom du site / service",
      baseline = "baseline - precisions sur l organisation",
      class = "fr-m-1w"
    ),
    title = "Exemple",
    fluidRow_dsfr(
      numericInput_dsfr(
        inputId = "mynumericinput",
        label = "Numeric input",
        value = 12,
        min = 0,
        max = NA,
        step = 2
      ),
      verbatimTextOutput(outputId = "numericinputvalue")
    )
  )
  server <- function(input, output, session) {
    output$numericinputvalue <- renderText({
      paste("La valeur de l'input est", input$mynumericinput)
    })
  }

  shinyApp(ui, server)
}
}
\seealso{
\code{\link[shiny:numericInput]{shiny::numericInput()}} in Shiny
}
