% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiobuttons_dsfr.R
\name{radioButtons_dsfr}
\alias{radioButtons_dsfr}
\title{radioButtons_dsfr}
\usage{
radioButtons_dsfr(inputId, label, choices, inline = FALSE, class = NULL)
}
\arguments{
\item{inputId}{id de l'input}

\item{label}{label du bouton}

\item{choices}{Liste des valeurs à sélectionner (si les éléments de la liste portent un nom, c'est ce nom qui est affiché à l'utilisateur et non la valeur)}

\item{inline}{Si TRUE, positionne les choix en ligne (c'est-à-dire horizontalement).}

\item{class}{des classes a ajouter si necessaire}
}
\value{
html
}
\description{
radioButtons_dsfr
}
\examples{
radioButtons_dsfr(
      inputId = "test", label = "Test", choices = c("A","B"), class = NULL
)

# radioButtons_dsfr(
#       inputId = "test", 
#       label = "Test", 
#       choices = c("Choix A" = "A","Choix B" = "B"),
#       class = NULL
# )
}
