% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modaldialog_dsfr.R
\name{removeModal_dsfr}
\alias{removeModal_dsfr}
\title{Faire dispparaitre une modale}
\usage{
removeModal_dsfr(
  inputId = "fr-modal-1",
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{inputId}{TODO}

\item{session}{TODO}
}
\description{
Faire dispparaitre une modale
}
