% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectinput_dsfr.R
\name{selectInput_dsfr}
\alias{selectInput_dsfr}
\title{selectInput_dsfr}
\usage{
selectInput_dsfr(inputId, label, choices, selected = NULL)
}
\arguments{
\item{inputId}{id de l'input}

\item{label}{label du bouton}

\item{choices}{Liste des valeurs à sélectionner (si les éléments de la liste portent un nom, c'est ce nom qui est affiché à l'utilisateur et non la valeur)}

\item{selected}{Element selectionné}
}
\value{
html
}
\description{
selectInput_dsfr
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {
  library(shiny)
  library(shinygouv)

  ui <- fluidPage_dsfr(
    header = header_dsfr(
      intitule = "Intitule",
      officiel = "Officiel",
      nom_site_service = "Nom du site / service",
      baseline = "baseline - precisions sur l organisation",
      class = "fr-m-1w"
    ),
    title = "Exemple",
    fluidRow_dsfr(
      # sans vecteur nommé
      selectInput_dsfr(
        inputId = "myselectInput",
        label = "Mon label",
        choices = c(
          "cyl",
          "am",
          "gear"
        )
      ),
      fluidRow_dsfr(
        # avec un vecteur nommé et une valeur sélectionnée
        selectInput_dsfr(
          inputId = "myselectInput2",
          label = "Mon label2",
          choices = c(
            "Cylinders" = "cyl",
            "Transmission" = "am",
            "Gears" = "gear"
          ),
          selected = "gear"
        )
      )
    )
  )
  server <- function(input, output, session) {
  }

  shinyApp(ui, server)
}
}
