% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinygouv-package.R
\docType{package}
\name{shinygouv-package}
\alias{shinygouv}
\alias{shinygouv-package}
\title{shinygouv: An Amazing Shiny App}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Juliette ENGELARE-LEFEBVRE \email{juliette.engelaere-lefebvre@developpement-durable.gouv.fr}

Authors:
\itemize{
  \item Sébastien Rochette \email{sebastien@thinkr.fr} (\href{https://orcid.org/0000-0002-1565-9313}{ORCID})
  \item Murielle Delmotte \email{murielle@thinkr.fr} (\href{https://orcid.org/0000-0002-1339-2424}{ORCID})
  \item Cervan Girard \email{cervan@thinkr.fr} (\href{https://orcid.org/0000-0002-4816-4624}{ORCID})
}

Other contributors:
\itemize{
  \item David Granjon [contributor]
  \item Alan Dipert [contributor]
}

}
\keyword{internal}
