% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modaldialog_dsfr.R
\name{showModal_dsfr}
\alias{showModal_dsfr}
\title{Faire apparaitre une modale}
\usage{
showModal_dsfr(
  ui,
  inputId = "fr-modal-1",
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{ui}{TODO}

\item{inputId}{TODO}

\item{session}{TODO}
}
\description{
Faire apparaitre une modale
}
\seealso{
\code{\link[shiny:showModal]{shiny::showModal()}} in Shiny
}
