% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatenumericinput_dsfr.R
\name{updateNumericInput_dsfr}
\alias{updateNumericInput_dsfr}
\title{updateSelectInput_dsfr}
\usage{
updateNumericInput_dsfr(
  inputId,
  label = NULL,
  value = NULL,
  min = NULL,
  max = NULL,
  step = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{inputId}{id de l'input}

\item{label}{label du bouton}

\item{session}{la session, la valeur par défaut est getDefaultReactiveDomain().}

\item{choices}{Liste des valeurs à sélectionner (si les éléments de la liste portent un nom, c'est ce nom qui est affiché à l'utilisateur et non la valeur)}

\item{selected}{Element selectionné}
}
\value{
html
}
\description{
updateSelectInput_dsfr
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {
  library(shiny)
  library(shinygouv)

  ui <- fluidPage_dsfr(
    header = header_dsfr(
      intitule = "Intitule",
      officiel = "Officiel",
      nom_site_service = "Nom du site / service",
      baseline = "baseline - precisions sur l organisation",
      class = "fr-m-1w"
    ),
    title = "Exemple",
    fluidRow_dsfr(
      numericInput_dsfr(
        inputId = "mynumericinput",
        label = "Numeric input",
        value = 12,
        min = 0,
        max = NA,
        step = 2
      ),
      actionButton_dsfr(
        inputId = "updateLabel",
        label = "updateLabel"
      ),
      actionButton_dsfr(
        inputId = "updateValue",
        label = "updateValue"
      ),
      actionButton_dsfr(
        inputId = "updateMin",
        label = "updateMin à la valeur en cours"
      ),
      actionButton_dsfr(
        inputId = "updateMax",
        label = "updateMax à la valeur en cours"
      ),
      actionButton_dsfr(
        inputId = "updateStep",
        label = "updateStep"
      ),
      verbatimTextOutput(outputId = "numericinputvalue")
    )
  )
  server <- function(input, output, session) {
    output$numericinputvalue <- renderText({
      paste("La valeur de l'input est", input$mynumericinput)
    })

    r <- reactiveValues()

    observeEvent(input$mynumericinput, {
      r$current_value <- input$mynumericinput
    })

    observeEvent(input$updateLabel, {
      updateNumericInput_dsfr(
        session = session,
        inputId = "mynumericinput",
        label = sample(LETTERS, 10)
      )
    })

    observeEvent(input$updateValue, {
      updateNumericInput_dsfr(
        session = session,
        inputId = "mynumericinput",
        value = round(
          runif(n = 1, min = 0, max = 1000)
        )
      )
    })

    observeEvent(input$updateMin, {
      updateNumericInput_dsfr(
        session = session,
        inputId = "mynumericinput",
        min = r$current_value
      )
    })

    observeEvent(input$updateMax, {
      updateNumericInput_dsfr(
        session = session,
        inputId = "mynumericinput",
        max = r$current_value
      )
    })

    observeEvent(input$updateStep, {
      updateNumericInput_dsfr(
        session = session,
        inputId = "mynumericinput",
        step = round(runif(n = 1, min = 0.1, max = 1), digits = 1)
      )
    })
  }

  shinyApp(ui, server)
}
}
