% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateradiogroupbuttons_dsfr.R
\name{updateRadioGroupButtons_dsfr}
\alias{updateRadioGroupButtons_dsfr}
\title{updateRadioGroupButtons_dsfr}
\usage{
updateRadioGroupButtons_dsfr(
  inputId,
  label = NULL,
  choices = NULL,
  selected = NULL,
  inline = FALSE,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{inputId}{id de l'input}

\item{label}{label du bouton}

\item{choices}{Liste des valeurs à sélectionner (si les éléments de la liste portent un nom, c'est ce nom qui est affiché à l'utilisateur et non la valeur)}

\item{selected}{Element selectionné}

\item{session}{la session, la valeur par défaut est getDefaultReactiveDomain().}
}
\value{
html
}
\description{
updateRadioGroupButtons_dsfr
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {
ui <- fluidPage_dsfr(
  radioGroupButtons_dsfr(inputId = "inradioGroupButtons", 
                    label = "Input radio buttons",
                    choices = c("Item A","Item B", "Item C")),
  actionButton_dsfr("go","Change")
)

server <- function(input, output, session) {
  
  observeEvent(input$go,{
    updateRadioGroupButtons_dsfr(
      session = session, 
      inputId = "inradioGroupButtons",
      label = "Un nouveau label",
      choices = c("A" = "a"),
      selected = "a"
    )
  })
  
  observeEvent(input$inradioGroupButtons,{
    message(input$inradioGroupButtons)
  })
}

shinyApp(ui, server)
}

}
