% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatetabsetpanel_dsfr.R
\name{updateTabsetPanel_dsfr}
\alias{updateTabsetPanel_dsfr}
\title{updateTabsetPanel_dsfr}
\usage{
updateTabsetPanel_dsfr(
  inputId,
  selected = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{inputId}{inputId du tabsetPanel_dsfr}

\item{selected}{titre du tab à afficher (paramètre \code{title} utilisé dans \code{tabPanel_dsfr()})}

\item{session}{session shiny}
}
\value{
html
}
\description{
updateTabsetPanel_dsfr
}
\examples{
if (interactive()){
  library(shiny)
  library(shinygouv)
  shinyApp(
    ui = fluidPage_dsfr(
    fluidRow_dsfr(
      column_dsfr(
        12,
        h3("Demo tabsetPanel_dsfr() et tabPanel_dsfr()"),
        # Adding space to the column
        # https://www.systeme-de-design.gouv.fr/elements-d-interface/fondamentaux-techniques/espacement
        extra_class = "fr-my-6w"
      ),
      column_dsfr(
        12,
        tabsetPanel_dsfr(
          "testtabpanel",
          tabPanel_dsfr(
            id = "tab1",
            title = "Une prems tab",
            content = tagList(
              h3("Coucou le tab 1")
            )
          ),
          tabPanel_dsfr(
            id = "tab2",
            title = "Une deuxieme tab",
            content = tagList(
              h3("Salut le tab 2")
            )
          )
        )
      )
    ),
    fluidRow_dsfr(
      column_dsfr(
        12,
        actionButton_dsfr(inputId = "showfirsttab",
                     "Afficher le premier onglet"
        ),
        actionButton_dsfr(inputId = "showsecondtab",
                     "Afficher le deuxi\u00e8me onglet"
        )
      )
    )
    ),
    server = function(input, output, session) {
          observeEvent(input$showfirsttab, {
      updateTabsetPanel_dsfr(
        inputId = "testtabpanel",
        selected = "Une prems tab",
        session = session
      )
    })

    observeEvent(input$showsecondtab, {
      updateTabsetPanel_dsfr(
        inputId = "testtabpanel",
        selected = "Une deuxieme tab",
        session = session
      )
    })
    }
  )
}
}
\seealso{
\href{https://shiny.posit.co/r/reference/shiny/latest/updatetabsetpanel}{updatetabsetpanel() in Shiny}
}
