% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withspinner_dsfr.R
\name{withSpinner_dsfr}
\alias{withSpinner_dsfr}
\title{withSpinner_dsfr}
\usage{
withSpinner_dsfr(expr, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{expr}{L'opération a réaliser.}

\item{session}{}
}
\value{
html
}
\description{
withSpinner_dsfr
}
\examples{
\dontrun{
# To be added in your app_server.R ou some server function on a module
withSpinner_dsfr(
  expr = {
    message("Je vais m'afficher pendant 5 secondes")
    Sys.sleep(5)
  }
)
}
}
